refactor: remove src/helpers folder
renamed: src/helpers/Html.js -> src/html.js
deleted: src/helpers/with-layout.js
