+++ src/express/middlewares/render-html.js
@@ -3,7 +3,7 @@ import { ServerStyleSheet, StyleSheetManager } from 'styled-components'
 import { StaticRouter } from 'react-router-dom'
 import App from '../../app'
 import get from 'lodash/get'
-import Html from '../../helpers/Html'
+import Html from '../../html'
 import Loadable from 'react-loadable'
 import React from 'react'
 import ReactDOMServer from 'react-dom/server'
